chore:chart admin fixes and cart deletion after payment

+++ backend/controllers/webhookController.js
@@ -5,6 +5,7 @@ import Order from "../models/orderModel.js";
 import Cart from "../models/cartModel.js";
 
 export const handlePaystackWebhook = async (req, res) => {
+      const userId = req.userId;
       //verify the request
       const secret = process.env.PAYSTACK_SECRET_KEY;
       const hash = crypto.createHmac("sha512", secret).update(JSON.stringify(req.body)).digest("hex");
@@ -47,12 +48,11 @@ export const handlePaystackWebhook = async (req, res) => {
                                     type: "charge",
                               },
                         );
-
+                        //update order status
                         await Order.findOneAndUpdate({ _id: order }, { paymentStatus: "paid", status: "processing" });
-                        const user = req.userId;
 
                         // Clear user's cart after verifying payment
-                        await Cart.findOneAndDelete({ user });
+                        await Cart.findOneAndDelete({ user: userId });
 
                         if (payment) console.log(`Payment ${payment._id} paid successfully.`);
                   }
